[로또] 4단계 - 로또(수동) (#4114)
* docs 요구사항 작성

* feat 수동 입력받는 기능 구현

* feat 출력 수정

* feat 수동+자동으로 로또 생성

* feat 구매비용 부족 예외처리 추가

* feat 컨트롤러에 예외처리 추가

* docs 문서 수정

* refactor 함수명 수정

* refactor LottoRequest 변수 final

* refactor 오타 수정

* refactor Money 객체 생성

* refactor 줄 정리

* test LottoRequestTest TC 정리

* refactor 로또 번호 생성 로직 수정

---------

+++ src/test/java/lotto/model/LottoResultTest.java
@@ -22,7 +22,7 @@ class LottoResultTest {
                         LottoPrize.NONE
                 )
         );
-        int profit = lottoResult.calculateProfit();
+        long profit = lottoResult.calculateProfit();
 
         assertThat(profit).isEqualTo(2_000_000_000 * 1 + 30_000_000 * 2 + 1_500_000 * 1 + 50_000 * 1 + 5_000 * 2 + 0 * 1);
     }
